+++ src/test/kotlin/racingcar/controller/WorldTest.kt
@@ -1,7 +1,8 @@
-package racingcar
+package racingcar.controller
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
+import racingcar.model.Car
 
 class WorldTest {
